test: refactor watch cwd tests
